requirements.txt
fauxmo==0.5.0
Flask==1.1.1
inkyphat==1.0.1
Pillow==6.1.0
